# Switch to megaparsec 6 for all versions of GHC