--- a/hledger-lib/Hledger/Read/Common.hs
+++ b/hledger-lib/Hledger/Read/Common.hs
@@ -112,9 +112,10 @@ import Data.Text (Text)
 import qualified Data.Text as T
 import Data.Time.Calendar
 import Data.Time.LocalTime
+import Data.Void (Void)
 import System.Time (getClockTime)
-import Text.Megaparsec.Compat
-import Control.Applicative.Combinators (skipManyTill)
+import Text.Megaparsec
+import Text.Megaparsec.Char
 
 import Hledger.Data
 import Hledger.Utils
@@ -181,13 +182,13 @@ rawOptsToInputOpts rawopts = InputOpts{
 --- * parsing utilities
 
 -- | Run a string parser with no state in the identity monad.
-runTextParser, rtp :: TextParser Identity a -> Text -> Either (ParseError Char MPErr) a
+runTextParser, rtp :: TextParser Identity a -> Text -> Either (ParseError Char Void) a
 runTextParser p t =  runParser p "" t
 rtp = runTextParser
 
 -- XXX odd, why doesn't this take a JournalParser ?
 -- | Run a journal parser with a null journal-parsing state.
-runJournalParser, rjp :: Monad m => TextParser m a -> Text -> m (Either (ParseError Char MPErr) a)
+runJournalParser, rjp :: Monad m => TextParser m a -> Text -> m (Either (ParseError Char Void) a)
 runJournalParser p t = runParserT p "" t
 rjp = runJournalParser
 
@@ -913,7 +914,7 @@ followingcommentandtagsp mdefdate = do
       runTextParser (setPosition pos *> parser) txt
 
     tagDate :: (SourcePos, Tag)
-            -> Either (ParseError Char MPErr) (TagName, Day)
+            -> Either (ParseError Char Void) (TagName, Day)
     tagDate (pos, (name, value)) =
       fmap (name,) $ runTextParserAt (datep' myear) (pos, value)
       where myear = fmap (first3 . toGregorian) mdefdate