--- a/hledger-lib/Hledger/Read/TimeclockReader.hs
+++ b/hledger-lib/Hledger/Read/TimeclockReader.hs
@@ -60,7 +60,8 @@ import           Data.Maybe (fromMaybe)
 import           Data.Text (Text)
 import qualified Data.Text as T
 import           Test.HUnit
-import           Text.Megaparsec.Compat hiding (parse)
+import           Text.Megaparsec hiding (parse)
+import           Text.Megaparsec.Char
 
 import           Hledger.Data
 -- XXX too much reuse ?