… hledger-lib/Text/Megaparsec/Compat.hs
… hledger-lib/package.yaml
… hledger-ui/hledger-ui.cabal
… hledger-ui/package.yaml
… hledger-web/Handler/AddForm.hs
… hledger-web/hledger-web.cabal
… hledger-web/package.yaml
… hledger/Hledger/Cli/CliOptions.hs
… hledger/Hledger/Cli/Commands/Add.hs
… hledger/hledger.cabal
… hledger/package.yaml
… stack-ghc7.10.yaml