This PR addresses #783 and
- Configures stack to use `megaparsec-6.4.1` for all versions of GHC we support
- Obviates and removes the megaparsec compatibility module `/hledger-lib/Text/Megaparsec/Compat.hs`
- Refactors parsers to use the more efficient `takeWhileP` (where I could find opportunity to do so)
- Refactors `accountnamep`

The parser is not much faster after these changes, but the memory usage has dropped significantly. I am guessing this is because `takeWhileP` allows us to obtain `Text` directly (as opposed to going through `String`) and `Text` is completely strict.

Before:
```
> hledger -f examples/10000x1000x10.journal register +RTS -s > /dev/null
   5,751,330,864 bytes allocated in the heap
     283,563,056 bytes copied during GC
      33,930,368 bytes maximum residency (10 sample(s))
         906,112 bytes maximum slop
              90 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0      5565 colls,     0 par    0.286s   0.287s     0.0001s    0.0017s
  Gen  1        10 colls,     0 par    0.169s   0.222s     0.0222s    0.0666s

  TASKS: 4 (1 bound, 3 peak workers (3 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.001s  (  0.001s elapsed)
  MUT     time    3.467s  (  3.506s elapsed)
  GC      time    0.455s  (  0.508s elapsed)
  EXIT    time    0.000s  (  0.006s elapsed)
  Total   time    3.924s  (  4.021s elapsed)

  Alloc rate    1,658,856,710 bytes per MUT second

  Productivity  88.4% of total user, 87.3% of total elapsed
```
After:
```
> hledger -f examples/10000x1000x10.journal register +RTS -s > /dev/null
   5,556,836,424 bytes allocated in the heap
     249,067,432 bytes copied during GC
      27,899,544 bytes maximum residency (10 sample(s))
         848,256 bytes maximum slop
              74 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0      5379 colls,     0 par    0.262s   0.262s     0.0000s    0.0013s
  Gen  1        10 colls,     0 par    0.167s   0.200s     0.0200s    0.0550s

  TASKS: 4 (1 bound, 3 peak workers (3 total), using -N1)

  SPARKS: 0 (0 converted, 0 overflowed, 0 dud, 0 GC'd, 0 fizzled)

  INIT    time    0.001s  (  0.001s elapsed)
  MUT     time    3.416s  (  3.462s elapsed)
  GC      time    0.428s  (  0.463s elapsed)
  EXIT    time    0.000s  (  0.005s elapsed)
  Total   time    3.845s  (  3.931s elapsed)

  Alloc rate    1,626,898,582 bytes per MUT second

  Productivity  88.8% of total user, 88.2% of total elapsed
```