@simonmichael 2018-05-22
Your numbers seems like a ~3% drop in allocation and time and a near 20% drop in maximum residency. Nice!